+++ src/test/java/fibonacciTest/FibonacciTest.java
@@ -1,6 +1,6 @@
-package factorialTest;
+package fibonacciTest;
 
-import factorial.Factorial;
+import fibonacci.Fibonacci;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -16,23 +16,23 @@ import static org.junit.jupiter.api.Assertions.*;
     * factorial negative number
  **/
 
-public class FactorialTest {
-    private Factorial factorial;
+public class FibonacciTest {
+    private Fibonacci fibonacci;
 
     @BeforeEach
     public void setup(){
-        factorial = new Factorial();
+        fibonacci = new Fibonacci();
     }
 
     @AfterEach
     public void finish (){
-        factorial = null;
+        fibonacci = null;
     }
 
     @Test
     public void testComputeReturnOneIfTheNumberIsZero(){
         int expectedValue = 1;
-        int obtainedValue = factorial.compute(0);
+        int obtainedValue = fibonacci.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
@@ -40,7 +40,7 @@ public class FactorialTest {
     @Test
     public void testComputeReturnOneIfTheNumberIsOne(){
         int expectedValue = 1;
-        int obtainedValue = factorial.compute(1);
+        int obtainedValue = fibonacci.compute(1);
 
         assertEquals(expectedValue, obtainedValue);
     }
@@ -48,7 +48,7 @@ public class FactorialTest {
     @Test
     public void testComputeReturnTwoIfTheNumberIsTwo(){
         int expectedValue = 2;
-        int obtainedValue = factorial.compute(2);
+        int obtainedValue = fibonacci.compute(2);
 
         assertEquals(expectedValue, obtainedValue);
     }
@@ -56,7 +56,7 @@ public class FactorialTest {
     @Test
     public void testComputeReturnSixIfTheNumberIsThree(){
         int expectedValue = 6;
-        int obtainedValue = factorial.compute(3);
+        int obtainedValue = fibonacci.compute(3);
 
         assertEquals(expectedValue, obtainedValue);
     }
@@ -64,13 +64,13 @@ public class FactorialTest {
     @Test
     public void testComputeReturn720IfTheNumberIs6(){
         int expectedValue = 6;
-        int obtainedValue = factorial.compute(3);
+        int obtainedValue = fibonacci.compute(3);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void testComputeOfANegativeNumberRaiseAnException(){
-        assertThrows(RuntimeException.class, () -> factorial.compute(-1));
+        assertThrows(RuntimeException.class, () -> fibonacci.compute(-1));
     }
 }
